removing cycles
