+++ utils.js
@@ -6,7 +6,7 @@ import Constants from 'expo-constants';
 import * as Notifications from 'expo-notifications';
 import * as SecureStore from 'expo-secure-store';
 import { Planning } from './planning';
-import { version, initialLineNumber } from "./App"
+import { version, initialLineNumber } from "./global.js"
 import CountDown from 'react-native-countdown-component';
 
 class Liste {
